brainbot requirements.txt
Python 3.10

TensorFlow

Streamlit

Pillow

NumPy